+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - improved `String` manipulation speed
 - `#count_in` and `#scan_in` methods for `String` interaction
-- new predefined sets `CharacterSet::assigned`, `::surrogate`
+- new predefined sets `CharacterSet::any`, `::assigned`, `::surrogate`
 - conversion methods `#assigned_part`, `#valid_part`
 - section methods `#plane(n)`, `#ascii_part`, `#ascii_part?`, `#ascii_only?`, `#ascii_ratio`, `#astral_only?`